refactor(modules): move IO + thread impl under Sockets/Threads
