+++ src/Reseau.jl
@@ -34,65 +34,40 @@ include("common/device_random.jl")
 include("common/encoding.jl")
 include("common/system_info.jl")
 include("common/uuid.jl")
-include("common/condition_variable.jl")
-include("common/thread.jl")
-include("common/thread_shared.jl")
-include("common/task_scheduler.jl")
 include("common/common.jl")
 
-# --- io ---
-include("io/io.jl")
+# --- public submodules (native implementations) ---
+include("threads/threads.jl")
+include("sockets/sockets.jl")
 
-# Previously included from src/io/io.jl (same order preserved)
-include("io/tracing.jl")
-include("io/event_loop_types.jl")
-include("io/kqueue_event_loop_types.jl")
-include("io/epoll_event_loop_types.jl")
-include("io/iocp_event_loop_types.jl")
-include("io/event_loop.jl")
-include("io/kqueue_event_loop.jl")
-include("io/epoll_event_loop.jl")
-include("io/iocp_event_loop.jl")
-include("io/message_pool.jl")
-include("io/posix_socket_types.jl")
-include("io/apple_nw_socket_types.jl")
-include("io/winsock_socket_types.jl")
-include("io/socket.jl")
-include("io/posix_socket_impl.jl")
-include("io/winsock_socket.jl")
-include("io/winsock_init.jl")
-include("io/blocks_abi.jl")
-include("io/apple_nw_socket_impl.jl")
-include("io/channel.jl")
-include("io/statistics.jl")
-include("io/socket_channel_handler.jl")
-include("io/host_resolver.jl")
-include("io/retry_strategy.jl")
-include("io/stream.jl")
-include("io/pem.jl")
-include("io/shared_library.jl")
-include("io/pkcs11.jl")
-include("io/pki_utils.jl")
-include("io/pipe.jl")
-include("io/iocp_pipe.jl")
-include("io/future.jl")
-include("io/channel_bootstrap.jl")
+# Re-alias moved bindings back into the parent module so existing call-sites
+# (including downstream packages) can keep referencing `Reseau.*`.
+function _alias_module_bindings!(src_mod::Module)::Nothing
+    for name in names(src_mod; all = true, imported = false)
+        str = String(name)
+        startswith(str, "@") && continue
+        startswith(str, "#") && continue
+        if isdefined(@__MODULE__, name)
+            # Allow overwriting Base/Core imports (we want the moved implementation
+            # bindings), but never clobber bindings defined by this module.
+            owner = Base.binding_module(@__MODULE__, name)
+            owner === (@__MODULE__) && continue
+            (owner === Base || owner === Core) || continue
+        end
+        val = getfield(src_mod, name)
+        @eval const $(name) = $src_mod.$(name)
+    end
+    return nothing
+end
 
-# Previously included directly from src/Reseau.jl
-include("io/aws_byte_helpers.jl")
-include("io/crypto_primitives.jl")
-include("io/async_stream.jl")
-include("io/tls_channel_handler.jl")
-include("io/alpn_handler.jl")
+_alias_module_bindings!(Threads)
+_alias_module_bindings!(Sockets)
 
-# --- public submodules (thin wrappers / new surfaces) ---
+# --- thin wrappers / grouped surfaces ---
 include("EventLoops.jl")
-include("Sockets.jl")
-# Must be last: defining `Reseau.Threads` shadows `Base.Threads` within this module.
-include("Threads.jl")
 
 function __init__()
-    _init_os_thread_cfunc!()
+    Threads._init_os_thread_cfunc!()
     io_library_init()
 end
