+++ src/sockets/io/crypto_primitives.jl
