+++ src/sockets/io/host_resolver.jl
@@ -966,11 +966,11 @@ function _native_getaddrinfo(hostname::String; flags::Cint = Cint(0))::Vector{Tu
 end
 
 """
-    getalladdrinfo(hostname; flags=0) -> Vector{Tuple{String, Cint}}
+    getalladdrinfo_raw(hostname; flags=0) -> Vector{Tuple{String, Cint}}
 
 Public wrapper for the libuv-free `getaddrinfo()` implementation used by the
 host resolver. This exists primarily so downstream packages can avoid `Sockets`.
 """
-function getalladdrinfo(hostname::AbstractString; flags::Cint = Cint(0))::Vector{Tuple{String, Cint}}
+function getalladdrinfo_raw(hostname::AbstractString; flags::Cint = Cint(0))::Vector{Tuple{String, Cint}}
     return _native_getaddrinfo(String(hostname); flags = flags)
 end
