+++ src/sockets/io/retry_strategy.jl
